requirements.txt
openai==1.99.3
python-dotenv==1.1.1
pandas==2.2.3
joblib==1.5.1
flask==3.1.2
requests==2.32.3
scikit-learn==1.7.1
joblib==1.5.1